Move inSphere into own MRInSphere.h/.cpp, document deviations from full SoS
The predicate shares no code with other precise functions and will be extended
in the following PRs. The header now lists all scenarios where the current
deterministic answers deviate from full simulation-of-simplicity (coincident
triangle points with a perturbation-creatable sphere, rSq exactly equal to the
squared circumradius), and the new sosInSphereDeviations test pins the current
behavior of each, together with the degenerate answers that are provably exact
(collinear triangle, third point at or beyond the diameter, rSq below the
squared circumradius).
